Configure: don't try to be clever when configuring afalgeng
If the afalgeng is enabled then Configure tries to be clever but fails, by only
actually building afalgeng if it isn't being cross-compiled and if the current
kernel is 4.1+.

This means that everyone cross compiling, or whose builder machines are running
older kernels (not uncommon for build farms), can't enable afalgeng.

Instead remove the cleverness and simply enable/disable as requested.

Fixes #7687